+++ .github/workflows/release.yml
@@ -15,7 +15,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        package_name: [clock, timer, stopwatch, dekamoji, components, style-bulma]
+        package_name: [clock, timer, stopwatch, dekamoji, measure, treemap, components, style-bulma]
 
     steps:
       - uses: actions/checkout@v3
