+++ packages/treemap/README.md
@@ -1,0 +1,11 @@
+# @kitsuyui/react-treemap
+
+## Demo
+
+Storybook: https://react-playground.storybook.kitsuyui.com/
+
+## Installation
+
+## License
+
+MIT
