+++ packages/treemap/jest.config.ts
@@ -1,0 +1,19 @@
+// https://jestjs.io/docs/configuration
+export default {
+  extensionsToTreatAsEsm: ['.ts', '.tsx'],
+  testPathIgnorePatterns: ['/node_modules/', '/dist/'],
+  testEnvironment: 'jsdom',
+  transform: {
+    '^.+.(t|j)sx?$': [
+      '@swc/jest',
+      {
+        sourceMaps: true,
+        jsc: {
+          parser: {
+            syntax: 'typescript',
+          },
+        },
+      },
+    ],
+  },
+}
