+++ packages/treemap/src/index.ts
@@ -1,0 +1,1 @@
+export { Treemap } from './treemap'
