+++ packages/treemap/tsup.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  target: 'es2020',
+  format: ['cjs', 'esm'],
+  entry: [
+    './src/**/*.{ts,tsx}',
+    '!./src/**/*.test.{ts,tsx}',
+    '!./src/**/test.{ts,tsx}',
+  ],
+  clean: true,
+  dts: true,
+})
